Refactored layout logic
